Finished ProcessInvite page. Need to make custom InviteLogin page or view
