+++ Areas/Identity/Pages/Account/Register.cshtml.cs
@@ -1,123 +1,123 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Text.Encodings.Web;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authorization;
-using BugTracker.Models;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI.Services;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.Extensions.Logging;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.ComponentModel.DataAnnotations;
+//using System.Linq;
+//using System.Text;
+//using System.Text.Encodings.Web;
+//using System.Threading.Tasks;
+//using Microsoft.AspNetCore.Authentication;
+//using Microsoft.AspNetCore.Authorization;
+//using BugTracker.Models;
+//using Microsoft.AspNetCore.Identity;
+//using Microsoft.AspNetCore.Identity.UI.Services;
+//using Microsoft.AspNetCore.Mvc;
+//using Microsoft.AspNetCore.Mvc.RazorPages;
+//using Microsoft.AspNetCore.WebUtilities;
+//using Microsoft.Extensions.Logging;
 
-namespace BugTracker.Areas.Identity.Pages.Account
-{
-    [AllowAnonymous]
-    public class RegisterModel : PageModel
-    {
-        private readonly SignInManager<BTUser> _signInManager;
-        private readonly UserManager<BTUser> _userManager;
-        private readonly ILogger<RegisterModel> _logger;
-        private readonly IEmailSender _emailSender;
+//namespace BugTracker.Areas.Identity.Pages.Account
+//{
+//    [AllowAnonymous]
+//    public class RegisterModel : PageModel
+//    {
+//        private readonly SignInManager<BTUser> _signInManager;
+//        private readonly UserManager<BTUser> _userManager;
+//        private readonly ILogger<RegisterModel> _logger;
+//        private readonly IEmailSender _emailSender;
 
-        public RegisterModel(
-            UserManager<BTUser> userManager,
-            SignInManager<BTUser> signInManager,
-            ILogger<RegisterModel> logger,
-            IEmailSender emailSender)
-        {
-            _userManager = userManager;
-            _signInManager = signInManager;
-            _logger = logger;
-            _emailSender = emailSender;
-        }
+//        public RegisterModel(
+//            UserManager<BTUser> userManager,
+//            SignInManager<BTUser> signInManager,
+//            ILogger<RegisterModel> logger,
+//            IEmailSender emailSender)
+//        {
+//            _userManager = userManager;
+//            _signInManager = signInManager;
+//            _logger = logger;
+//            _emailSender = emailSender;
+//        }
 
-        [BindProperty]
-        public InputModel Input { get; set; }
+//        [BindProperty]
+//        public InputModel Input { get; set; }
 
-        public string ReturnUrl { get; set; }
+//        public string ReturnUrl { get; set; }
 
-        public IList<AuthenticationScheme> ExternalLogins { get; set; }
+//        public IList<AuthenticationScheme> ExternalLogins { get; set; }
 
-        public class InputModel
-        {
-            [Required]
-            [EmailAddress]
-            [Display(Name = "Email")]
-            public string Email { get; set; }
+//        public class InputModel
+//        {
+//            [Required]
+//            [EmailAddress]
+//            [Display(Name = "Email")]
+//            public string Email { get; set; }
 
-            [Required]
-            [Display(Name = "First Name")]
-            public string FirstName { get; set; }
+//            [Required]
+//            [Display(Name = "First Name")]
+//            public string FirstName { get; set; }
 
-            [Required]
-            [Display(Name = "Last Name")]
-            public string LastName { get; set; }
+//            [Required]
+//            [Display(Name = "Last Name")]
+//            public string LastName { get; set; }
 
-            [Required]
-            [StringLength(100, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
-            [DataType(DataType.Password)]
-            [Display(Name = "Password")]
-            public string Password { get; set; }
+//            [Required]
+//            [StringLength(100, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
+//            [DataType(DataType.Password)]
+//            [Display(Name = "Password")]
+//            public string Password { get; set; }
 
-            [DataType(DataType.Password)]
-            [Display(Name = "Confirm password")]
-            [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
-            public string ConfirmPassword { get; set; }
-        }
+//            [DataType(DataType.Password)]
+//            [Display(Name = "Confirm password")]
+//            [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
+//            public string ConfirmPassword { get; set; }
+//        }
 
-        public async Task OnGetAsync(string returnUrl = null)
-        {
-            ReturnUrl = returnUrl;
-            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
-        }
+//        public async Task OnGetAsync(string returnUrl = null)
+//        {
+//            ReturnUrl = returnUrl;
+//            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
+//        }
 
-        public async Task<IActionResult> OnPostAsync(string returnUrl = null)
-        {
-            returnUrl ??= Url.Content("~/");
-            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
-            if (ModelState.IsValid)
-            {
-                var user = new BTUser { UserName = Input.Email, Email = Input.Email, FirstName = Input.FirstName, LastName = Input.LastName };
-                var result = await _userManager.CreateAsync(user, Input.Password);
-                if (result.Succeeded)
-                {
-                    _logger.LogInformation("User created a new account with password.");
+//        public async Task<IActionResult> OnPostAsync(string returnUrl = null)
+//        {
+//            returnUrl ??= Url.Content("~/");
+//            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
+//            if (ModelState.IsValid)
+//            {
+//                var user = new BTUser { UserName = Input.Email, Email = Input.Email, FirstName = Input.FirstName, LastName = Input.LastName };
+//                var result = await _userManager.CreateAsync(user, Input.Password);
+//                if (result.Succeeded)
+//                {
+//                    _logger.LogInformation("User created a new account with password.");
 
-                    var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-                    code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-                    var callbackUrl = Url.Page(
-                        "/Account/ConfirmEmail",
-                        pageHandler: null,
-                        values: new { area = "Identity", userId = user.Id, code = code, returnUrl = returnUrl },
-                        protocol: Request.Scheme);
+//                    var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+//                    code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
+//                    var callbackUrl = Url.Page(
+//                        "/Account/ConfirmEmail",
+//                        pageHandler: null,
+//                        values: new { area = "Identity", userId = user.Id, code = code, returnUrl = returnUrl },
+//                        protocol: Request.Scheme);
 
-                    await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
-                        $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
+//                    await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
+//                        $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
 
-                    if (_userManager.Options.SignIn.RequireConfirmedAccount)
-                    {
-                        return RedirectToPage("RegisterConfirmation", new { email = Input.Email, returnUrl = returnUrl });
-                    }
-                    else
-                    {
-                        await _signInManager.SignInAsync(user, isPersistent: false);
-                        return LocalRedirect(returnUrl);
-                    }
-                }
-                foreach (var error in result.Errors)
-                {
-                    ModelState.AddModelError(string.Empty, error.Description);
-                }
-            }
+//                    if (_userManager.Options.SignIn.RequireConfirmedAccount)
+//                    {
+//                        return RedirectToPage("RegisterConfirmation", new { email = Input.Email, returnUrl = returnUrl });
+//                    }
+//                    else
+//                    {
+//                        await _signInManager.SignInAsync(user, isPersistent: false);
+//                        return LocalRedirect(returnUrl);
+//                    }
+//                }
+//                foreach (var error in result.Errors)
+//                {
+//                    ModelState.AddModelError(string.Empty, error.Description);
+//                }
+//            }
 
-            // If we got this far, something failed, redisplay form
-            return Page();
-        }
-    }
-}
+//            // If we got this far, something failed, redisplay form
+//            return Page();
+//        }
+//    }
+//}
