+++ Areas/Identity/Pages/Account/RegisterByInvite.cshtml
@@ -1,8 +1,8 @@
-﻿@*@page
+﻿@page 
 @model RegisterModel
 @{
-    ViewData["Title"] = "Register";
-    ViewData["MainMenu"] = "Register";
+    ViewData["Title"] = "Register by Invite";
+    ViewData["MainMenu"] = "Register by Invite";
     Layout = "~/Views/Shared/_LayoutAuth.cshtml";
 }
 
@@ -49,4 +49,3 @@
 <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
 <script src="~/popper.min.js"></script>
 <script src="~/bootstrap/js/bootstrap.min.js"></script>  
-*@
